move 'event_start' and 'event_end' out of site_state
These are simple functions we're using a lot; having them in site_state
(which has a lot of dependencies) caused some circular imports.

This also tidies up usage of `event_start().year` and replaces it with
`event_year()`.
