+++ models/site_state.py
@@ -29,14 +29,6 @@ VALID_STATES = {
 }
 
 
-def event_start():
-    return config_date("EVENT_START")
-
-
-def event_end():
-    return config_date("EVENT_END")
-
-
 def calc_site_state(date):
     """ Logic to set the state of the homepage based on date. """
     if date < config_date("SALES_START"):
